+++ include/ozz/base/io/archive.h
@@ -370,6 +370,16 @@ OZZ_INLINE const internal::Array<const _Ty> MakeArray(const _Ty* _array,
   const internal::Array<const _Ty> array = {_array, _count};
   return array;
 }
+template <typename _Ty>
+OZZ_INLINE const internal::Array<_Ty> MakeArray(Range<_Ty> _array) {
+  const internal::Array<_Ty> array = {_array.begin, _array.Count()};
+  return array;
+}
+template <typename _Ty>
+OZZ_INLINE const internal::Array<const _Ty> MakeArray(Range<const _Ty> _array) {
+  const internal::Array<const _Ty> array = {_array.begin, _array.Count()};
+  return array;
+}
 template <typename _Ty, size_t _count>
 OZZ_INLINE const internal::Array<_Ty> MakeArray(_Ty (&_array)[_count]) {
   const internal::Array<_Ty> array = {_array, _count};