+++ test/base/io/archive_tests.cc
@@ -147,6 +147,8 @@ TEST(PrimitiveArrays, Archive) {
     o << ozz::io::MakeArray(fo);
     const uint32_t* po_null = NULL;
     o << ozz::io::MakeArray(po_null, 0);
+    const ozz::Range<const float> rfo(fo);
+    o << ozz::io::MakeArray(rfo);
 
     // Read primitive types.
     stream.Seek(0, ozz::io::Stream::kSet);
@@ -183,6 +185,10 @@ TEST(PrimitiveArrays, Archive) {
     EXPECT_EQ(std::memcmp(fi, fo, sizeof(fo)), 0);
     uint32_t* pi_null = NULL;
     i >> ozz::io::MakeArray(pi_null, 0);
+    float fi2[OZZ_ARRAY_SIZE(fo)];
+    ozz::Range<float> rfi(fi2);
+    i >> ozz::io::MakeArray(rfi);
+    EXPECT_EQ(std::memcmp(rfi.begin, fo, sizeof(fo)), 0);
   }
 }